Fix broken category and tag links
Since Hugo 0.120.0 the output of `delimit` is now deHTMLized. For
example, it turns '<a>' into '&lt;a&gt;' in the final output. `delimit`
is used to create lists of categories and tags in the blog posts
summary, so it affects these lists, breaking HTML links in them.

We can bring it back to working HTML by piping it to the `safeHTML`
function. This commit does that.