Add visual diff job
This will comment on a PR if it causes a visual change in the example
page when using the theme version in the PR compared to the one on the
master branch. The idea is to discover unintentional changes to the
appeareance.

The workflow relies on playwright for serving and taking screenshots of
the example website before and after the change in the PR. And it relies
on pixelmatch to compare the images. Pixelmatch displays a faded version
of the original screenshot with changes highligted in red (antialiasing
changes are highligted in yellow).